# oresight-backend/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OreSight API",
    "description": "Backend API for OreSight - a mine reserve intelligence prototype built for SIH26009 (MOIL manganese mining).",
    "version": "0.1.0"
  },
  "paths": {
...
    "/sites/{site_id}": {
      "get": {
        "tags": [
          "sites"
        ],
        "summary": "Get a single mine site",
        "description": "Return one mine site by id, or 404 if it doesn't exist.",
        "operationId": "get_site_sites__site_id__get",
        "parameters": [
          {
            "name": "site_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Site Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SiteOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/sites/{site_id}/geojson": {
      "get": {
        "tags": [
          "sites"
        ],
        "summary": "Get a site's boundary as GeoJSON",
        "description": "Return the site's boundary polygon as a GeoJSON FeatureCollection, ready\nfor direct use as a MapLibre GeoJSON source.",
        "operationId": "get_site_geojson_sites__site_id__geojson_get",
        "parameters": [
          {
            "name": "site_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Site Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A GeoJSON FeatureCollection with a single Feature for the site polygon",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Site Geojson Sites  Site Id  Geojson Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SiteOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "belt_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Belt Name"
          },
          "district": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "District"
          },
          "state": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "State"
          },
          "centroid_lat": {
            "type": "number",
            "title": "Centroid Lat"
          },
          "centroid_lon": {
            "type": "number",
            "title": "Centroid Lon"
          },
          "active_risk_count": {
            "type": "integer",
            "title": "Active Risk Count"
          },
          "avg_reserve_confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avg Reserve Confidence"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "belt_name",
          "district",
          "state",
          "centroid_lat",
          "centroid_lon",
          "active_risk_count",
          "avg_reserve_confidence"
        ],
        "title": "SiteOut",
        "example": {
          "active_risk_count": 2,
          "avg_reserve_confidence": 0.67,
          "belt_name": "Balaghat Manganese Belt",
          "centroid_lat": 21.8,
          "centroid_lon": 80.19,
          "district": "Balaghat",
          "id": 1,
          "name": "Balaghat",
          "state": "Madhya Pradesh"
        }
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}